requirements.txt
asgiref==3.8.1
Django==5.2.3
django-browser-reload==1.12.1
django-tailwind==4.0.1
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
djangorestframework==3.16.0
django-simple-history==3.10.1
django-import-export==4.3.8
tablib==3.7.0
django-filter==24.1

django-crispy-forms==2.4
crispy-daisyui==0.7.0
pillow==11.3.0

openpyxl==3.1.5
django-weasyprint==2.4.0
geopy==2.4.1
requests==2.32.4
django-tables2==2.7.5